% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioReport-package.R
\docType{package}
\name{PhysioReport-package}
\alias{PhysioReport}
\alias{PhysioReport-package}
\title{PhysioReport: clinical report generation for the x-biosignal ecosystem}
\description{
Parameterized, bilingual (JA/EN), colorblind-safe clinical reports with
normative overlays and MDC/MCID-annotated change.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/x-biosignal/PhysioReport}
  \item \url{https://x-biosignal.github.io/PhysioReport/}
  \item Report bugs at \url{https://github.com/x-biosignal/PhysioExperiment/issues}
}

}
\author{
\strong{Maintainer}: Yusuke Matsui \email{mail.to.matsui@gmail.com}

}
\keyword{internal}
